requirements.txt
typer>=0.12,<1.0
telethon>=1.36,<2.0
PyYAML>=6.0,<7.0
python-dateutil>=2.9,<3.0
rich>=13,<14
anthropic>=0.40,<1.0
